requirements.txt
streamlit>=1.30.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
google-genai>=0.1.0